move aria-label to link
