+++ packages/ui/src/components/ResourceBlock/BaseItem.tsx
@@ -1,10 +1,7 @@
 import { ReactNode } from "react";
 import CheckCircleIcon from "@heroicons/react/20/solid/CheckCircleIcon";
 import ExclamationCircleIcon from "@heroicons/react/20/solid/ExclamationCircleIcon";
-import { useIntl } from "react-intl";
 import { tv } from "tailwind-variants";
-
-import { commonMessages } from "@gc-digital-talent/i18n";
 
 const stateIcon = tv({
   base: "absolute top-5 right-5 size-5",
@@ -37,53 +34,31 @@ const baseItem = tv({
 
 export interface BaseItemProps {
   title: ReactNode;
-  accessibleLabel: string;
   description: ReactNode;
   state?: "incomplete" | "complete";
 }
 
-const BaseItem = ({
-  title,
-  accessibleLabel,
-  description,
-  state,
-}: BaseItemProps) => {
-  const intl = useIntl();
-  let combinedLabel;
-  switch (state) {
-    case "incomplete":
-      combinedLabel = `${accessibleLabel} (${intl.formatMessage(commonMessages.incomplete)})`;
-      break;
-    case "complete":
-      combinedLabel = `${accessibleLabel} (${intl.formatMessage(commonMessages.complete)})`;
-      break;
-    default:
-      combinedLabel = accessibleLabel;
-  }
-
-  return (
+const BaseItem = ({ title, description, state }: BaseItemProps) => (
+  <div
+    className={baseItem({
+      class:
+        state === "incomplete"
+          ? // should match the absolute positioning of the center of the state icon (x0.75 + (x0.75/2))
+            "bg-radial-[circle_7.5rem_at_top_1.6875rem_right_1.6875rem] from-error/10 to-transparent"
+          : undefined,
+    })}
+    role="listitem"
+  >
+    <StateIcon state={state} />
     <div
-      className={baseItem({
-        class:
-          state === "incomplete"
-            ? // should match the absolute positioning of the center of the state icon (x0.75 + (x0.75/2))
-              "bg-radial-[circle_7.5rem_at_top_1.6875rem_right_1.6875rem] from-error/10 to-transparent"
-            : undefined,
-      })}
-      aria-label={combinedLabel}
-      role="listitem"
+      // icon extends margin + icons size: x0.75 + x0.75 = x1.5
+      // item margin is base(x1) l-tablet(x1.5)
+      className="pr-3"
     >
-      <StateIcon state={state} />
-      <div
-        // icon extends margin + icons size: x0.75 + x0.75 = x1.5
-        // item margin is base(x1) l-tablet(x1.5)
-        className="pr-3"
-      >
-        {title}
-      </div>
-      <p className="text-sm text-gray-600 dark:text-gray-200">{description}</p>
+      {title}
     </div>
-  );
-};
+    <p className="text-sm text-gray-600 dark:text-gray-200">{description}</p>
+  </div>
+);
 
 export default BaseItem;
